requirements.txt
anyio==3.6.1
click==8.1.3
Faker==15.0.0
fastapi==0.85.0
graphql-core==3.2.3
greenlet==1.1.3
h11==0.13.0
httptools==0.5.0
idna==3.4
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
strawberry-graphql==0.133.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
